service/user/urls.py
from django.urls import path
from rest_framework.permissions import AllowAny
from rest_framework_simplejwt.views import (
   TokenObtainPairView,
   TokenRefreshView,
)


from user.apps import UserConfig
from user.views import UserCreateAPIView

app_name = UserConfig.name


urlpatterns = [
    path('login/', TokenObtainPairView.as_view(permission_classes =(AllowAny,)), name='login'),
    path('token/refresh/', TokenRefreshView.as_view(permission_classes =(AllowAny,)), name='token_refresh'),
    path('register/', UserCreateAPIView.as_view(), name='register'),
]
